src/agent/dspy_modules.py
"""
DSPy modules for optimizing prompts and enforcing constraints.

Phase 4 implementation: Use DSPy for hypothesis relevance and optimization.
"""

try:
    import dspy
    from dspy import Assert, Suggest
    DSPY_AVAILABLE = True
except ImportError:
    DSPY_AVAILABLE = False
    print("Warning: DSPy not available. Install with: pip install dspy-ai")


if DSPY_AVAILABLE:
    class HypothesisFormation(dspy.Module):
        """DSPy module for hypothesis formation with constraints."""
        
        def __init__(self):
            super().__init__()
            self.generator = dspy.ChainOfThought(
                "query, sub_question, observations -> hypothesis, confidence, reasoning"
            )
        
        def forward(self, query, sub_question, observations):
            """
            Generate hypothesis with assertions and suggestions.
            
            Enforces:
            - Hypothesis must relate to sub-question
            - Confidence should be proportional to evidence
            """
            result = self.generator(
                query=query,
                sub_question=sub_question,
                observations=observations
            )
            
            # Enforce: hypothesis must relate to sub-question
            dspy.Assert(
                self._relates_to_subquestion(result.hypothesis, sub_question),
                "Hypothesis doesn't address the sub-question"
            )
            
            # Guide: confidence should correlate with evidence
            num_obs = len(observations) if isinstance(observations, list) else 1
            expected_max_confidence = min(1.0, num_obs / 10.0)
            
            dspy.Suggest(
                result.confidence <= expected_max_confidence,
                f"Confidence seems high given only {num_obs} observations"
            )
            
            return result
        
        def _relates_to_subquestion(self, hypothesis: str, sub_question: str) -> bool:
            """Check if hypothesis relates to sub-question."""
            # Simple heuristic: check for keyword overlap
            hyp_words = set(hypothesis.lower().split())
            sq_words = set(sub_question.lower().split())
            
            # Remove common words
            common = {'the', 'a', 'an', 'is', 'are', 'was', 'were', 'what', 'who', 'when', 'where', 'why', 'how'}
            hyp_words -= common
            sq_words -= common
            
            overlap = len(hyp_words & sq_words)
            return overlap >= 2  # At least 2 keywords in common
    
    
    class MetaReasoning(dspy.Module):
        """DSPy module for meta-reasoning with constraints."""
        
        def __init__(self):
            super().__init__()
            self.reasoner = dspy.ChainOfThought(
                "query, sub_questions, hypotheses, iterations -> can_answer, confidence, decision, reasoning"
            )
        
        def forward(self, query, sub_questions, hypotheses, iterations, max_iterations):
            """
            Meta-reasoning with assertions about decision quality.
            """
            result = self.reasoner(
                query=query,
                sub_questions=sub_questions,
                hypotheses=hypotheses,
                iterations=iterations
            )
            
            # Assert: decision must be valid
            valid_decisions = ['CONTINUE', 'STOP', 'SYNTHESIZE']
            dspy.Assert(
                result.decision in valid_decisions,
                f"Decision must be one of {valid_decisions}"
            )
            
            # Suggest: high confidence should lead to synthesis
            if result.confidence >= 8.0:
                dspy.Suggest(
                    result.decision == 'SYNTHESIZE',
                    "High confidence should lead to synthesis"
                )
            
            # Suggest: max iterations should stop
            if iterations >= max_iterations:
                dspy.Suggest(
                    result.decision in ['STOP', 'SYNTHESIZE'],
                    "Max iterations reached, should stop or synthesize"
                )
            
            return result
    
    
    def optimize_agent_prompts(agent, training_data, metric_fn):
        """
        Optimize agent prompts using DSPy.
        
        Args:
            agent: ECU agent instance
            training_data: List of (query, gold_answer) pairs
            metric_fn: Function to evaluate answer quality
            
        Returns:
            Optimized agent
        """
        from dspy.teleprompt import MIPROv2
        
        optimizer = MIPROv2(
            metric=metric_fn,
            auto='light',
            num_candidates=5,
        )
        
        optimized = optimizer.compile(
            agent,
            trainset=training_data,
        )
        
        return optimized


else:
    # Dummy implementations if DSPy not available
    
    class HypothesisFormation:
        def __init__(self):
            pass
        
        def forward(self, query, sub_question, observations):
            return {"hypothesis": "", "confidence": 0.0, "reasoning": ""}
    
    class MetaReasoning:
        def __init__(self):
            pass
        
        def forward(self, query, sub_questions, hypotheses, iterations, max_iterations):
            return {"can_answer": False, "confidence": 0.0, "decision": "CONTINUE", "reasoning": ""}
    
    def optimize_agent_prompts(agent, training_data, metric_fn):
        print("DSPy not available, skipping optimization")
        return agent

